% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnp.R
\name{coef.multinomial_probit}
\alias{coef.multinomial_probit}
\title{Extract multinomial probit model parameters}
\usage{
\method{coef}{multinomial_probit}(object, ...)
}
\arguments{
\item{object}{A multinomial probit regression model as output by function \link{multinomial.probit}.}

\item{...}{Not used.}
}
\value{
A list with two entries:\itemize{
\item `coef` Containing the fitted coefficients, as a matrix of dimensions [n_features + intercep, n_classes - 1].
The first level of `y` is used as the base class and thus has no coefficients (could be thought of as being all
zeros).
\item `Sigma` Containing the fitted class covariance matrix, as a square matrix of dimension `n_classes`.
The entry in the first row and first column is always fixed to 1.
}
}
\description{
Extract the estimated coefficients and covariance from a
multinomial probit regression model.
}
