% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnp.R
\name{mnp.likelihood}
\alias{mnp.likelihood}
\title{Multinomial Probit Log-Likelihood}
\usage{
mnp.likelihood(model, x, y, weights = NULL, nthreads = parallel::detectCores())
}
\arguments{
\item{model}{A multinomial probit regression model as output by function \link{multinomial.probit}.}

\item{x}{The covariates. If it is passed as a `data.frame`, factor columns should have
the same levels as the `x` data to which the model was fitted. Note that this package
does not perform column matching - the number of columns must match exactly with the
`x` data to which the model was fitted.}

\item{y}{The labels. Should be passed as a factor variable, with the same levels as the
`y` to which the model was fitted.}

\item{weights}{Optional row weights.}

\item{nthreads}{Number of parallel threads to use.}
}
\value{
The obtained log-likelihood (a negative number, with higher values meaning a better
math between predictions and labels).
}
\description{
Calculates the log-likelihood on a given set of covariates and labels
from a fitted multinomial probit model.
}
