% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnp.R
\name{multinomial.probit}
\alias{multinomial.probit}
\title{Multinomial Probit Regression}
\usage{
multinomial.probit(
  x,
  y,
  weights = NULL,
  lambda = 0,
  intercept = TRUE,
  grad = c("analytical", "finite_diff"),
  presolve_logistic = TRUE,
  nthreads = parallel::detectCores()
)
}
\arguments{
\item{x}{The covariates. If passing a `data.frame`, will be encoded through `model.matrix`, and the user is responsible
for handling the levels of factor variables.}

\item{y}{The labels. Should be passed as a `factor` type.}

\item{weights}{Optional row weights.}

\item{lambda}{Amount of L2 regularization. Note that the regularization is added following GLMNET's formula.}

\item{intercept}{Whether to add intercepts to the coefficients for each class.}

\item{grad}{How to calculate gradients of the MVN CDF parameters.

If passing "analytical", will use the theoretically correct formula for the gradients given by
the MVN PDF plus conditioned CDF (using Plackett's identity for the correlation coefficient
gradients), with the same CDF approximation as used for calculating the likelihood.

If passing "finite_diff", will approximate the gradients through finite differencing.

In theory, analytical gradients should be more accurate and faster, but in practive, since
the CDF is calculated by an imprecise approximation, the analytical gradients of the theoretical
CDF might not be too accurate as gradients of this approximation. Both approaches involve the
same number of CDF calculations, but for analytical gradients, the problems are reduced by
one / two dimensions, which makes them slightly faster.

Note that this refers to gradients w.r.t the parameters of MVN distributions, not w.r.t to the
actual model parameters, which are still obtained by applying the chain rule.

On smaller datasets in particular, finite differencing can result in more accurate gradients,
but on larger datasets, as errors average out, analytical gradients tend to be more accurate.}

\item{presolve_logistic}{Whether to pre-solve for the coefficients by fitting a multinomial \bold{logistic} regression
first and use that solution as starting point for optimization (after a smaller second
presolve pass solving only for the Cholesky of the covariance matrix).}

\item{nthreads}{Number of parallel threads to use.}
}
\value{
The fitted model object (class "multinomial_probit"), on which methods such as
\link{predict.multinomial_probit} or \link{coef.multinomial_probit} can be called.
}
\description{
Fits a multinomial probit regression (for multi-class classification with 3 or more classes), with
likelihoods and gradients computed through a fast approximation to the CDF of the MVN distribution
based on the TVBS method (see references for details) and optimized through the BFGS method.

In general, the approximation for the likelihood and gradients is not good enough for the resulting
model to be very usable - typically, results are not competitive against simpler multinomial logistic
regression (which assumes a unit covariance matrix), and as such, is not recommended for serious usage.
}
\details{
If the data is a sparse matrix, it is highly recommended to load library
"MatrixExtra" before running this function, and to pass the "x" data as a CSR matrix of
class "dgRMatrix".

Note that this library uses OpenMP for parallelization. As such, it is recommended to use
BLAS / LAPACK libraries that are OpenMP aware, such as `openblas-openmp` (which is usually
not the default openblas variant) or MKL (you might also want to set environment variable
`MKL_THREADING_LAYER=GNU`).
}
\examples{
library(multinomial.probit)
data(iris)
x <- iris[, 1:4]
y <- iris$Species
m <- multinomial.probit(x, y)
predict(m, x, type="prob")
predict(m, x, type="class")
}
\references{
\enumerate{
\item Bhat, Chandra R.
"New matrix-based methods for the analytic evaluation of the multivariate cumulative normal distribution function."
Transportation Research Part B: Methodological 109 (2018): 238-256.
\item Plackett, Robin L. "A reduction formula for normal multivariate integrals." Biometrika 41.3/4 (1954): 351-360.
}
}
