% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnp.R
\name{predict.multinomial_probit}
\alias{predict.multinomial_probit}
\title{Prediction function for multinomial probit model}
\usage{
\method{predict}{multinomial_probit}(
  object,
  newdata,
  type = c("prob", "class", "logprob"),
  nthreads = parallel::detectCores()
)
}
\arguments{
\item{object}{A multinomial probit regression model as output by function \link{multinomial.probit}.}

\item{newdata}{The new covariates for which to make predictions.}

\item{type}{The type of prediction to output. Options are:\itemize{
\item `"prob"` Will return the probabilities of each row belonging to each class (returned as a matrix).
\item `"class"` Will return the class with the highest predicted probability (returned as a factor vector).
\item `"logprob"` Will return log-probabilities for each row and class (returned as a matrix).
}}

\item{nthreads}{Number of parallel threads to use.}
}
\value{
Either a matrix or a vector depending on argument `type`.
}
\description{
Calculate predictions from a multinomial probit
model on new data.
}
\details{
The calculation of log-probabilities by the model uses a not-so-accurate
approximation, thus the outputs for each row are rescaled to sum up
exactly to one.
}
