% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnp.R
\name{print.multinomial_probit}
\alias{print.multinomial_probit}
\title{Print method for multinomial probit models}
\usage{
\method{print}{multinomial_probit}(x, ...)
}
\arguments{
\item{x}{A multinomial probit regression model as output by function \link{multinomial.probit}.}

\item{...}{Not used.}
}
\value{
The same `x` argument, as invisible.
}
\description{
Displays summary information from a multinomial probit
regression model.
}
