% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnp.R
\name{summary.multinomial_probit}
\alias{summary.multinomial_probit}
\title{Summary method for multinomial probit models}
\usage{
\method{summary}{multinomial_probit}(object, ...)
}
\arguments{
\item{object}{A multinomial probit regression model as output by function \link{multinomial.probit}.}

\item{...}{Not used.}
}
\value{
The same `object` argument, as invisible.
}
\description{
Displays summary information from a multinomial probit
regression model (same as method `print`).
}
